add valida cliente interface and tests
