+++ src/test/java/classes/PessoaTest.java
@@ -1,0 +1,19 @@
+package classes;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class PessoaTest{
+
+    private Pessoa pessoa;
+
+    @BeforeEach
+    public void setUp() throws Exception {
+        pessoa = new Pessoa();
+    }
+
+
+}
